+++ autrainer/datasets/utils/target_transforms/abstract_target_transform.py
@@ -1,0 +1,68 @@
+from abc import ABC, abstractmethod
+from typing import Any, List, Union
+
+import audobject
+import torch
+
+
+class AbstractTargetTransform(ABC, audobject.Object):
+    def __init__(self) -> None:
+        """Abstract target transform for handling target or label
+        transformations in a dataset.
+
+        Serves as the base for creating custom target transforms that handle
+        encoding and decoding targets or labels of a dataset, obtaining
+        predictions from a batch of model outputs, and determining the majority
+        vote from a list of targets.
+        """
+
+    def __call__(self, x: Any) -> Any:
+        return self.encode(x)
+
+    @abstractmethod
+    def encode(self, x: Any) -> Any:
+        """Encode a target or label.
+
+        Args:
+            x: Target or label.
+
+        Returns:
+            Encoded target or label.
+        """
+
+    @abstractmethod
+    def decode(self, x: Any) -> Any:
+        """Decode a target or label. Serve as the inverse operation of encode.
+
+        Args:
+            x: Encoded target or label.
+
+        Returns:
+            Decoded target or label.
+        """
+
+    @abstractmethod
+    def predict_batch(self, x: torch.Tensor) -> Union[List[Any], Any]:
+        """Get encoded predictions from a batch of model outputs.
+
+        Args:
+            x: Batch of model outputs.
+
+        Returns:
+            Encoded predictions.
+        """
+
+    @abstractmethod
+    def majority_vote(self, x: List[Any]) -> Any:
+        """Get the majority vote from a list of decoded targets or labels.
+
+        The majority vote is defined by the subclasses and may be the most
+        frequent target, the average of all targets, or any other operation
+        that determines a majority vote based on the list of targets or labels.
+
+        Args:
+            x: List of decoded targets or labels.
+
+        Returns:
+            Decoded majority vote.
+        """
